Allow failure on unknown fields
This change allows the ObjectMapper to be configured to fail when the response
contains an unknown field. This feature should normally be turned off, but it
can be useful in testing scenarios where we want to verify we can read all of
the fields in the response.

This change also allows the new unit tests to optionally check whether we know
about all of the fields in the payload. This will make it easier to keep up with
additions to the API in the future.