Warn if no User-Agent string provided
This change warns the user if no User-Agent string is provided. Chess.com
now requires all PubAPI requests to have valid contact information in the
User-Agent header. From my local testing, it seems that they are not
currently rejecting requests, so for now we will still attempt the request,
but we will also log a warning with a link to the relevant information on
Chess.com's website so as to encourage developers to start setting this
field.